requirements.txt
tweepy==4.14.0
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
nltk==3.8.1
python-dotenv==1.0.0
jupyter==1.0.0
notebook==7.0.3
selenium==4.11.2
webdriver-manager==4.0.0 